:sparkles: Implemented hook for custom camunda plugins
Refactored the hook to pass around the submission instance, which can
be used by custom plugins.

This implementation makes it possible to do blocking IO while obtaining
the reference, so you can poll the Camunda process for a particular
variable and even persist some results in the registration_result
JSONField of the submission, which may be exposed to the confirmation
template in the future.

This unlocks a lot of custom behaviour depending on how you use Camunda.